Tests/LibWeb/Text/expected/ServiceWorker/service-worker-get-registrations.txt

getRegistrations is a function: true
getRegistrations resolves with an array: true
getRegistrations resolves with a frozen array: true
getRegistrations resolves with an empty array: true
